[SQLite] copy tables during restore instead of replacing entire database
Summary:
[ENG-10489](https://linear.app/comm/issue/ENG-10489/copy-tables-instead-of-replacing-entire-database).

The Online Backup API completely erases data in the existing (target database). This causes the user to be logged out because olm_persist_account is not included in the backup, which means when we download and restore, the Olm account is completely erased, which causes CSAT invalidation.

Depends on D14518

Test Plan: Existing tables (e.g. `olm_persist_account`) are not overwritten during restore

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14519