[web] Disable all the actions for a deleted message
Summary:
When a message is deleted, we shouldn't be able to perform any action to it, e.g. edit, react, pin, etc. On native, it is handled automatically because `DeletedMessage` doesn't handle a click action. On web, we have to set the tooltip actions to an empty array.

https://linear.app/comm/issue/ENG-10494/update-reaction-logic

{F5353475}

Depends on D14546

Test Plan: Checked on both native and web that it isn't possible to e.g. react to a deleted message.

Reviewers: kamil, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14547