[native][web] Hide edit label and reactions of deleted messages
Summary: When a message is deleted, we should hide all the inline engagement components but a sidebars - because it is possible that a sidebar was created before the message was deleted.

https://linear.app/comm/issue/ENG-10494/update-reaction-logic
https://linear.app/comm/issue/ENG-10493/update-sidebar-creation-logic

Depends on D14547

Test Plan: Checked on both native and web that reactions and edit label aren't shown for a deleted message.

Reviewers: kamil, ashoat