[lib] Don't show a deleted message in a preview
Summary:
Solve the issue in the most general way, where a message spec doesn't need to be aware of being deletable. We're checking messages in an order, which means we can build a set of deleted messages as we go. Then simply continue if a message was deleted by one of the already seen messages.

https://linear.app/comm/issue/ENG-10462/make-sure-that-deleted-messages-are-displayed-correctly-in-the-preview

Depends on D14550

Test Plan: Send two messages, delete the most recent one and check if the older one is displayed in the preview.

Reviewers: kamil, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14551