[keyserver] Handle deleted messages when searching
Summary:
When we're creating a new message, we're updating a search index. This means that when we delete a message, we can delete all the index entries associated with the deleted message.

https://linear.app/comm/issue/ENG-10517/consider-deleted-message-in-message-search

Depends on D14552

Test Plan: Created a message, deleted it, sent 20 more messages, changed active thread, refreshed the app, opened the previous active thread, and searched for the message. After this diff - mo messages are shown. Before this diff - the original message content was shown (if less than 20 messages were sent, the `deleted message` was shown in the results).

Reviewers: kamil, ashoat

Differential Revision: https://phab.comm.dev/D14556