[native][web] Handle deleted messages when searching
Summary:
Delete from a search index when deleting a message.

https://linear.app/comm/issue/ENG-10517/consider-deleted-message-in-message-search

Depends on D14556

Test Plan: In a thick thread: created a message, deleted it, sent 20 more messages, changed active thread, refreshed the app, opened the previous active thread, and searched for the message. After this diff - mo messages are shown. Before this diff - the original message content was shown (if less than 20 messages were sent, the `deleted message` was shown in the results).

Reviewers: kamil, ashoat