[SQLite][native] introduce `SQLiteUtils` class
Summary:
There are some utility stateless functions in `SQLiteQueryExecutor`, that are used in various methods. This makes it impossible to move them (methods that use utilities) out of `SQLiteQueryExecutor`.

Because of this change here, we'll have more flexibility and we'll make things cleaner.

[ENG-10538](https://linear.app/comm/issue/ENG-10538/create-sqlite-utils-class).

Test Plan: Build native app, just moving code around

Reviewers: bartek, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14570