[native] Fix uploading images with ph-upload:// scheme
Summary:
https://linear.app/comm/issue/ENG-10531/unable-to-send-screenshots-in-latest-build

Related: D14512

Some images and videos on iOS have ph-upload:// scheme and they fail to be encrypted thus crashing the app. They work ok for processing though.
So we still use uri from image picker for processing but use local uri for encryption and upload.

Test Plan:
Try to upload a couple of images on iOS: a screenshot, an image from a camera, a video from a camera etc.
Do the same for Android

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14566