[lib] Move sending multimedia messages to the protocol
Summary:
Move the logic. The only changes were about moving some types around and changing thread infos with a single thread info.

https://linear.app/comm/issue/ENG-10592/move-message-sending-logic-to-specs

Depends on D14596

Test Plan: Run the app and check if it is possible to send a multimedia message on both native and web in thin and thick threads.

Reviewers: kamil, angelika

Reviewed By: angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14593