[lib] Modify change thread settings type to include thread info
Summary:
To make the logic compatible with the specs, we need to have thread type for both thin and thick threads. We can't unify the types further because thread changes are different depending on thread thickness - this has to be addressed later while introducing a new protocol.

https://linear.app/comm/issue/ENG-10636/move-managing-members-logic-to-specs

Depends on D14598

Test Plan: Flow

Reviewers: kamil, angelika

Reviewed By: angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14599