[lib] Create a flag that tells if actions can target pending messages
Summary:
In thick threads, we allow creating a sidebar and reacting to a message that is pending. We do that because a message can be pending because it failed to be sent to some recipients, while succeeding to some others.

Depends on D14617

Test Plan: Disable the internet, send a message in a thick thread, and check if the reaction action is available. Do the same in a thin thread - this time it shouldn't be available.

Reviewers: kamil, angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14622