[lib] Add a flag telling if messages are stored on a server
Summary:
Make the logic more general - we aren't interested here in a distinction between thin and thick thread. What matters here is whether the source of truth for messages is local or remote, and if we can fetch messages if we want to.

Depends on D14631

Test Plan: Tested on native: create a thick thread with many messages. Close it, open, and scroll - all the messages should appear.

Reviewers: kamil, angelika