[native] Fix processed files disposing
Summary:
https://linear.app/comm/issue/ENG-10675/original-non-processed-media-is-attempted-to-be-disposed-after-upload

To reproduce send a png from media gallery that doesn't require processing on iOS.
In this case:
selection.uri is an asset uri.
uploadURI is a filesystem uri of the same file
As we're comparing those two and they're different, shouldDisposePath points to original file and it's attempted to be disposed.
However, only the file after processing should be disposed.

As the solution instead of a comparison that is unclear and prone to bugs, processImage and processVideo simply return what path
should be disposed. They know better if and what they processed.

Test Plan:
- send an image that does require processing
- send an image that doesn't require processing
- send an image from a camera
- send a video that does require processing
- send a video that doesn't require processing

and verify in logs that only the processed and encrypted (.dat) file is disposed and it's disposed successfully.

Reviewers: bartek, tomek, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14636