[lib] Move converting DB threads to a spec
Summary:
Move the logic. This required introducing a type parameter to the specs. This refactoring also revealed an issue with our validation here: due to Flow, we weren't strictly checking if e.g. thick member type is used in a thick thread. This diff keeps this issue, and it won't be easy to fix.

Depends on D14640

Test Plan: Check if restoring thin and thick threads from DB works.

Reviewers: kamil, angelika

Reviewed By: angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14641