[web] use `threadActivityStore` from SQLite
Summary:
Context in [ENG-10671](https://linear.app/comm/issue/ENG-10671/add-threadactivitystore-communitystore-and-keyserverstore-to-blacklist).

We return `threadActivityStore: state` but `threadActivityStore` on the web is not in `redux-persist` which means on each refresh, we clear `threadActivityStore` and never read the contents from SQLite

Depends on D14643

Test Plan: Refresh the app and check that `threadActivityStore` is not cleared.

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14650