[lib] fix `threadActivityStore` assertion code
Summary:
Context in [ENG-10671](https://linear.app/comm/issue/ENG-10671/add-threadactivitystore-communitystore-and-keyserverstore-to-blacklist).

`onStateDifference` was never passed to this reducer, which means even if the ops logic has a bug, we never saw any alert connected to this, as it was silenced.

Depends on D14650

Test Plan: Simulate state difference and make sure Alert is presented on native

Reviewers: tomek, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14651