[web] wait with `callGetInitialReduxState` until SQLite state is loaded
Summary:
Context in [ENG-10671](https://linear.app/comm/issue/ENG-10671/add-threadactivitystore-communitystore-and-keyserverstore-to-blacklist).

The issue is described in [this comment](https://linear.app/comm/issue/ENG-10671/add-threadactivitystore-communitystore-and-keyserverstore-to-blacklist#comment-ca0d3014)

Depends on D14653

Test Plan: Log the steps and make sure order is correct and each step is executed only once. Rest is tested in the next diff in the stack

Reviewers: tomek, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14655