[web] handle unsupported version in ErrorBoundary
Summary:
[ENG-10627](https://linear.app/comm/issue/ENG-10627/something-went-wrong-on-web-when-version-is-unsupported).

The error from task description was thrown in `InitialReduxStateGate` because client was unable to talk to keyserver. It is not possible to use `VersionUnsupportedModal` because error in `InitialReduxStateGate` causes `initialStateLoaded` to be false and as a result Modals are not shown and only `<Loading />` is visible.

This is more general approach that should handle all cases, not only errors from `InitialReduxStateGate`.

Also note, that this scenario shouldn't really happen, also while testing D14592. Code from `InitialReduxStateGate` is executed on app refresh, which means that user should be already running on a newest version. We should only handle this error when `MinCodeVersion` changes while app is already loaded and is communicating with keyserver - this is already covered in `MinVersionHandler`, or during login in `useHandleSecondaryDeviceLogInError`.

Test Plan: Similar to D14592, bump the version and reload the app

Reviewers: tomek, bartek