[web] handle unsupported version in ErrorBoundary
Summary:
[ENG-10627](https://linear.app/comm/issue/ENG-10627/something-went-wrong-on-web-when-version-is-unsupported).

The error from the task description was thrown in `InitialReduxStateGate` because the client was unable to talk to keyserver. It is not possible to use `VersionUnsupportedModal` because the error in `InitialReduxStateGate` causes `initialStateLoaded` to be false, and as a result, Modals are not shown and only `<Loading />` is visible.

This more general approach should handle all cases, not only errors from `InitialReduxStateGate`.

Also note that this scenario shouldn't really happen while testing D14592. Code from `InitialReduxStateGate` is executed on app refresh, which means that the user should already be running on the newest version. We should only handle this error when `MinCodeVersion` changes while the app is already loaded and is communicating with keyserver - this is already covered in `MinVersionHandler`, or during login in `useHandleSecondaryDeviceLogInError`.

Test Plan: Similar to D14592, bump the version and reload the app

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14657