[SQLite] move connection creation to manager class and remove `default_on_db_open_callback` and
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

`default_on_db_open_callback` was huge blocker to move some code around and to make some fields non-static. It was needed only on native, so it was moved to `NativeSQLiteConnectionManager` as `onDatabaseOpen`.

`initializeConnection` was converted to be pure virtual to provide platform specific implementation on each platform (they're different). To avoid code duplication `createConnection` was introduced, but note that this is only and util, caching `dbConnection` was moved to derived clasess.

Depends on D14662

Test Plan: Login/logout on web and native, perform some operations on db

Reviewers: bartek, tomek, angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14663