[lib] Generate composable message IDs deterministically from local IDs
Summary:
We were generating new IDs every time we tried to send composable messages - this was causing issues when we retried a message. To fix that, we're changing an approach so that a retry yields the same ID as the first attempt.

Note: this should've introduced a util function that transforms a local ID into the ID. Currently, this would introduce an import cycle. Adding this function can be safely performed in D14613.

https://linear.app/comm/issue/ENG-10746/retried-messages-have-different-ids-on-the-senders-and-on-the

Test Plan:
1. Log in on both native and web
2. Disable the internet on the web
3. Send a text message
4. Enable the internet
5. Retry the message
6. React on the message on both clients
The reactions should appear on both clients

Reviewers: kamil, angelika