[lib] avoid resetting keyserver's `updatesCurrentAsOf` on `setClientDBStoreActionType` action
Summary: [ENG-10762](https://linear.app/comm/issue/ENG-10762/keyserverstore-is-read-with-database-with-updatescurrentasof-0).

This condition is always triggerred on `setClientDBStoreActionType` because we populate empty store with store from DB, so `updatesCurrentAsOf` are different (0 vs BD value).

This was causing state sync on each app open after we removed presisting `keyserverStore` in `SQLite`.

Test Plan: There are no state sync on each app open

Reviewers: tomek