[lib] avoid resetting keyserver's `updatesCurrentAsOf` on `setClientDBStoreActionType` action
Summary:
[ENG-10762](https://linear.app/comm/issue/ENG-10762/keyserverstore-is-read-with-database-with-updatescurrentasof-0).

This condition is line 130 is always triggered on `setClientDBStoreActionType` because we populate the empty store with the store from DB, so `updatesCurrentAsOf` are different (0 vs database value).

This was causing state sync on each app open after we removed presisting `keyserverStore` in `SQLite`.

Test Plan: There are no state sync on each app open

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14690