[SQLite] fix and improve `generateBackupDataKey`
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

`generateBackupDataKey` was currently generating two keys, which is extremely misleading. In the next diffs, those methods are going to be moved to `DatabaseManager`, so first, a refactor is in place to avoid confusing reviewers.

Added some comments to make it clear.

Depends on D14716

Test Plan: Build the app and check that database is still there and can be used without logging out user. Then logout to check if generation still works.

Reviewers: bartek, tomek, angelika