[SQLite] move key generation to `DatabaseManager`
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

We need to move it higher to be able to create multiple `SQLiteQueryExecutor`'s. Logic should remain unchanged; later in the stack, there are more improvements to avoid having static props and to remove `SQLiteQueryExecutor::initialize`. Additionally, this code is not used on the web, so there is no need to keep it in a shared place

`initializeSQLiteQueryExecutorProperties` implementation is mostly a copy-paste of old SQLiteQueryExecutor::initialize.

Depends on D14717

Test Plan: Build the app and check that database is still there and can be used without logging out user. Then logout to check if generation still works.

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14718