[SQLite] temporary make `SQLiteQueryExecutor` methods and fields public
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

NOTE: This is temporary to make refactoring possible, next diffs could either not build or have incorrect logic without this. Later in the stack, this is updated to be private again.

Depends on D14719

Test Plan: Build the app

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14720