[SQLite] move `triggerBackupFileUpload` to `DatabaseManager`
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

It is not executing any queries, it is daabase management.

It should be const because it changes connectionManager
Note that this IS SAVE because I am only moving static to a different class.

Depends on TBD

Test Plan: Build app

Reviewers: bartek, tomek, angelika