[SQLite] make functions in `EntityQueryHelpers.h` inline
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

Because this is in `.h` file, it causes a `duplicate symbol` on native, even if we use `#pragma once` (because this fails on linking, not on the compilation step). This should fix the issue.

Alternative is to move implementation to `.cpp` file, but not sure if this is worth it, I think the compiler will optimize it anyway, and I don't think this is worse in terms of readability.

Depends on D14720

Test Plan: Run tests

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14721