[SQLite] make functions in `EntityQueryHelpers.h` inline
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

Because this is in `.h` file it causes `duplicate symbol` on native even if we use `#pragma once` (because this fails on linking, not on compilation step). This should fix the issue.

Depends on D14720

Test Plan: Run tests

Reviewers: bartek, tomek, angelika