[native] Respect the session overriding
Summary: Currently, we're ignoring the `override` flag which means that the native client will never accept the fact that it lost a race condition with another device. The native code should be similar to what we have for web in `contentInboundSessionCreator` inside `worker-crypto.js`.

Additionally, improve the logging a bit to catch additional errors.

https://linear.app/comm/issue/ENG-10754/olm-sessions-seem-to-break

Test Plan: It's hard to test this condition - run the app and sent a message to know if session creation isn't completely broken.

Reviewers: kamil, bartek, ashoat