[SQLite] make `connectionManager` to be local prop `SQLiteQueryExecutor` and move management to `DatabaseManager`
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

I decided to use `std::shared_ptr` because on native, we might have multiple `SQLiteQueryExecutor` objects using one connection - this is because objects are created with `thread_local`.

This is not great for multithreading, because multiple threads can at the same time modify `connectionManager`; however, we use a queue and `GlobalDBSingleton`, which should mitigate.
Also note that this is not worse than it was; in past, multiple threads could also concurrently modify the state of the class, because props were static.

There should be exactly one `static std::shared_ptr<NativeSQLiteConnectionManager> connectionManager;` per `thread_local SQLiteQueryExecutor`.

Previously, we were relying on the fact that `initializeQueryExecutor` is executed first and set the static props of `SQLiteQueryExecutor` - this is the same, the difference is that it sets the pointer. For web, we set this in constructor

Depends on D14733

Test Plan: restore, login/logout, sending some messages, creating drafts and checking if those are persisted

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14734