[SQLite] make `SQLiteQueryExecutor::getConnection` to be non-static
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

The goal is to make `connectionManager` non-static. We first need to convert the methods that use it.

`thread_local SQLiteQueryExecutor instance;` in DatabaseManager works, but it is not clean, but this is going to change in D14734.

Depends on D14731

Test Plan: Open the app and try to execute some queries

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14732