[SQLite] move starting log monitoring to `DatabaseManager`
Summary:
Implemented as part of [ENG-10603](https://linear.app/comm/issue/ENG-10603/biggest-risk-factor-out-code-from-databasequeryexecutor)

This is native-specific and should be placed in the shared `SQLiteQueryExecutor`. It should also be executed only once, as this is more of a `connectionManager` responsibility than `SQLiteQueryExecutor`—in the previous implementation, there was a bug; this code was executed as many times as many threads were using the database.

Depends on D14737

Test Plan: Flip the flag and check if logs are uploaded

Reviewers: bartek, tomek, angelika