[SQLite] make it possible to skip migrations in `SQLiteQueryExecutor` constructor
Summary:
[ENG-8358](https://linear.app/comm/issue/ENG-8358/run-the-migrations-during-backup-restoration)

For backup, we need to first apply logs before migrating, becuase of that we need to avoid running migrations in constructor. Alternative is factoring out migrate call but creating database work on this assumptions for years now and I don't want to change it.

Unfortunately, Emscripten is not great with default params, so I had to add it everywhere

Depends on D14743

Test Plan: Test creating instance without running migrations

Reviewers: bartek, tomek, angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14744