[web][native] create two `SQLiteQueryExecutor` instances
Summary:
[ENG-8358](https://linear.app/comm/issue/ENG-8358/run-the-migrations-during-backup-restoration)

Creating two instances and updating all the code around.

I was having some issues with `undefined symbol` on native when using default arguments, so I decided to overload the method.

Created [ENG-10786](https://linear.app/comm/issue/ENG-10786/database-downloaded-from-the-backup-is-not-encrypted) to track the fact that database is not encrypted

Depends on D14746

Test Plan: Tested in D14752

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14748