[native][web] update `processDBStoreOperations` to handle multiple databases
Summary:
[ENG-10611](https://linear.app/comm/issue/ENG-10611/native-support-databases-when-processing-ops) and [ENG-10610](https://linear.app/comm/issue/ENG-10610/web-support-databases-when-processing-ops)
We want to specify on which database to perform operations.

I on purpose made it mandatory in `SQLiteAPI`, to always force someone using API to consider which database to use, and optional to `commCoreModule`/worker to avoid modifications in old migrations code - note that we should always use `SQLiteAPI`, not the `commCoreModule`/worker directly.

Depends on D14753

Test Plan: Perform some operations on backup database before restoring

Reviewers: bartek, tomek, angelika

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14754