[native][RN79][skip-ci] Remove expo-secure-store
Summary:
https://linear.app/comm/issue/ENG-10284/migrate-expo-secure-store-related-code

There is a code in Podfile that says it's needed to compile expo secure store in Notification Service. I tried to remove it, but I was getting
errors like "sharedApplication not available for use in an App Extension". My guess is the notification service still compiles some RN stuff.

Depends on D14978

Test Plan: Build the app

Reviewers: bartek, ashoat

Reviewed By: bartek, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14979