[native][RN79][skip-ci] Update metro.config, .babelrc.cjs
Summary:
https://linear.app/comm/issue/ENG-10256/upgrade-react-native-to-080-and-expo-to-sdk-53

Before we were using `metro-react-native-babel-preset`, which was causing some errors when running the app.
Now we use `babel-preset-expo`, which extends react-native preset anyway and follows expo app structure.
I also turned off `unstable_enablePackageExports` to solve some issue with resolving packages.
The rest of the configuration is exactly as it was before.

Depends on D14991

Test Plan: Run the app

Reviewers: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14992