[native][web] add `plaintextDatabasePath` to `restoreFromMainCompaction`
Summary:
[ENG-10607](https://linear.app/comm/issue/ENG-10607/web-refactor-persisting-to-handle-more-sqlitequeryexecutor-objects)

Follow-up to https://phab.comm.dev/D14751#inline-81041.

This allows using predefined path to avoid need to persist it on web, we can use const as previously. This is also prerequisite to [ENG-10786](https://linear.app/comm/issue/ENG-10786/database-downloaded-from-the-backup-is-not-encrypted) because with this change, we can either decrypt database and use encryption on a different layer (web) or just avoid creating plaintext file (native).

Added comment to explain how it works.

Depends on D14755

Test Plan: Test restore on web

Reviewers: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14764