[lib] Confirm a message after processing it unconditionally
Summary:
We were confirming a message only if a condition is met, but there's no good reason to do so: if we were able to decrypt and process a message, we can confirm it. This should reduce the chance of receiving the same message again.

https://linear.app/comm/issue/ENG-10787/new-session-is-created-due-to-olm-already-decrypted-or-keys-skipped

Depends on D14756

Test Plan: Flow

Reviewers: kamil, angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14769