[lib][native][web] implement handler to sync store version
Summary:
Part of [ENG-7014](https://linear.app/comm/issue/ENG-10619/handle-case-when-backup-version-is-newer-than-the-app-version).

For primary device, we need to dispatch it to set this property in database, to later when restoring be able to check what migrations we need to run. Previously, it was set only during register which is not enough, we need to migrate all users to have this value in SQLite and this handler achieves it.

I added this on both platforms, because for secondary devices, this is helpful when after restoring backup was newer, with this we can easily detect using selectors when app was upgraded and we can easily attempt restore again.

It is save to rename it, previous name was confusing and not used currently, it was introduced for backup purposes but approach changed since.

Depends on D14788

Test Plan: Check if this is set in store and in database

Reviewers: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14789