[lib][native][web] implement handler to sync store version
Summary:
Part of [ENG-7014](https://linear.app/comm/issue/ENG-10619/handle-case-when-backup-version-is-newer-than-the-app-version).

For the primary device, we need to dispatch it to set this property in the database, to later, when restoring, be able to check what migrations we need to run. Previously, it was set only during registration, which is not enough; we need to migrate all users to have this value in SQLite, and this handler achieves it.

I added this on both platforms, because for secondary devices, this is helpful when after restoring backup was newer; with this, we can easily detect using selectors when the app was upgraded, and we can easily attempt to restore again.

It is safe to rename `db_version` -> `store_version`, the previous name was confusing and not used currently, it was introduced for backup purposes, but the approach has changed since.

Depends on D14788

Test Plan: Check if this is set in store and in database

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14789